Move SettableFuture to a different module
